refactor: apis/shops를 apis/shop으로 이름 변경
