+++ src/components/shop/register/ShopRegister.tsx
@@ -2,7 +2,7 @@ import { ChangeEvent, useState } from "react";
 import { useForm } from "react-hook-form";
 import * as z from "zod";
 
-import { postImages, postShopRegistData, putPresignedURL } from "@/apis/shops";
+import { postImages, postShopRegistData, putPresignedURL } from "@/apis/shop";
 import ShopDataForm from "@/components/shop/ShopDataForm";
 import { PAGE_ROUTES } from "@/routes";
